chore: split unit and integration  tests
