+++ README.md
@@ -26,7 +26,7 @@ Dirk van der Laarse
 
 #### Tests
 
-To run unit tests:
+To run unit and integration tests:
 
 ```shell
 bench --site test_site run-tests --app woocommerce_fusion --coverage
